+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 ### Added
 - Add `EnvelopeTemplateDocument.positions` endpoint
+- Add Suggest Contact EP `GET /api/my/contacts/suggest`
+- Add `MyEnvelopeRecipient` authenticationMethod properties
 - Add `EnvelopeDocument.fromTemplate`
 - Add `EnvelopeRecipient.fromTemplate`
 - Add `EnvelopeTag.fromTemplate`